refactoring

+++ src/main/java/com/parovsky/traver/service/impl/CategoryServiceImpl.java
@@ -1,12 +1,12 @@
 package com.parovsky.traver.service.impl;
 
 import com.parovsky.traver.dao.CategoryDAO;
+import com.parovsky.traver.dao.LocationDAO;
 import com.parovsky.traver.dto.model.SaveCategoryModel;
 import com.parovsky.traver.dto.model.UpdateCategoryModel;
 import com.parovsky.traver.dto.view.UserView;
 import com.parovsky.traver.entity.Category;
-import com.parovsky.traver.exception.EntityAlreadyExistsException;
-import com.parovsky.traver.exception.EntityNotFoundException;
+import com.parovsky.traver.exception.ApplicationException;
 import com.parovsky.traver.service.CategoryService;
 import com.parovsky.traver.service.UserService;
 import lombok.AllArgsConstructor;
@@ -14,13 +14,20 @@ import org.springframework.lang.NonNull;
 import org.springframework.stereotype.Service;
 
 import javax.validation.Valid;
+import java.util.Collections;
 import java.util.List;
+
+import static com.parovsky.traver.exception.Errors.*;
+import static com.parovsky.traver.utils.Constraints.ID;
+import static com.parovsky.traver.utils.Constraints.NAME;
 
 @Service
 @AllArgsConstructor(onConstructor = @__({@org.springframework.beans.factory.annotation.Autowired}))
 public class CategoryServiceImpl implements CategoryService {
 
 	private final CategoryDAO categoryDAO;
+
+	private final LocationDAO locationDAO;
 
 	private final UserService userService;
 
@@ -30,20 +37,20 @@ public class CategoryServiceImpl implements CategoryService {
 	}
 
 	@Override
-	public List<Category> getFavoriteCategories() throws EntityNotFoundException {
+	public List<Category> getFavoriteCategories() {
 		UserView user = userService.getCurrentUser();
 		return categoryDAO.getAllFavorite(user.getEmail());
 	}
 
 	@Override
-	public Category getCategoryById(@NonNull Long id) throws EntityNotFoundException {
-		return categoryDAO.get(id).orElseThrow(() -> new EntityNotFoundException("Category not found"));
+	public Category getCategoryById(@NonNull Long id) {
+		return categoryDAO.get(id).orElseThrow(() -> new ApplicationException(CATEGORY_NOT_FOUND, Collections.singletonMap(ID, id)));
 	}
 
 	@Override
-	public Category saveCategory(@NonNull SaveCategoryModel model) throws EntityAlreadyExistsException {
+	public Category saveCategory(@NonNull SaveCategoryModel model) {
 		if (categoryDAO.isExistByName(model.getName())) {
-			throw new EntityAlreadyExistsException("Category already exist");
+			throw new ApplicationException(CATEGORY_ALREADY_EXIST, Collections.singletonMap(NAME, model.getName()));
 		}
 		Category category = Category.builder()
 				.name(model.getName())
@@ -53,16 +60,19 @@ public class CategoryServiceImpl implements CategoryService {
 	}
 
 	@Override
-	public Category updateCategory(@Valid @NonNull UpdateCategoryModel model) throws EntityNotFoundException {
-		Category category = categoryDAO.get(model.getId()).orElseThrow(() -> new EntityNotFoundException("Category not found"));
+	public Category updateCategory(@Valid @NonNull UpdateCategoryModel model) {
+		Category category = categoryDAO.get(model.getId()).orElseThrow(() -> new ApplicationException(CATEGORY_NOT_FOUND, Collections.singletonMap(ID, model.getId())));
 		category.setName(model.getName());
 		category.setPicture(model.getPicture());
 		return categoryDAO.save(category);
 	}
 
 	@Override
-	public void deleteCategory(@NonNull Long id) throws EntityNotFoundException {
-		Category category = categoryDAO.get(id).orElseThrow(() -> new EntityNotFoundException("Category not found"));
+	public void deleteCategory(@NonNull Long id) {
+		Category category = categoryDAO.get(id).orElseThrow(() -> new ApplicationException(CATEGORY_NOT_FOUND, Collections.singletonMap(ID, id)));
+		if (locationDAO.isLocationExistsByCategoryId(id)) {
+			throw new ApplicationException(CATEGORY_HAS_LOCATIONS, Collections.singletonMap(ID, id));
+		}
 		categoryDAO.delete(category);
 	}
 }
